+++ test/backdrop_test.dart
@@ -1,4 +1,4 @@
-import 'package:backdrop/backdrop.dart';
+import 'package:backdrop_widget/backdrop.dart';
 import 'package:test/test.dart';
 
 void main() {